requirements.txt
chainlit

azure-ai-projects

azure-identity
